let DPKG_COLORS default to our APT::Color setting
dpkg can optionally colorize its output since 1.18.5. Currently this
defaults to 'never', but it will eventually be 'auto'. It seems
reasonable to assume that a user who has enabled/disabled colors in apt
will want to have dpkg have the same state regarding color usage.

This isn't overriding explicit settings by the user, so in case a user
feels strongly about it one way or the other there are options.