edsp: support generic and solver-specific configs
The spec was slightly inconsistent if the preferences setting is
available only as generic or specific setting & the code only supported
the specific one, while for the strict-pinning was only generic…

As the usual pattern for apt is to have both options we adapt the spec
and code to support both as well.

This also adds a purely informal "Solver" field so in case the request
is saved in a file, we know to which solver the sent preferences apply.

Closes: 823918