remove semi-support for different build-dirs
The debian/rules file tries to guess in which directory it is supposed
to be building, but that guess is always ./build – if it wasn't it
would fail later as not all rules take alternatives into acount.

So, as this is clearly not used lets remove this complexity instead of
fixing it up.

Git-Dch: Ignore