tests: avoid time-dependent rebuild of caches
The tests changes the sources.list and the modification time of this
file is considered while figuring out if the cache can be good. Usually
this isn't an issue, but in that case we have the cache generation
produce warnings which appear twice in this case.

(cherry picked from commit 70bef3257a4dc7751444db8dadedd207bd24ab35)