CMake: Improve handling of vendor files
First of all, instead of creating the files at configure time,
generate the files using normal target. This has the huge advantage
that they are rebuilt if their input changes. While we are at it,
also add dependencies on the vendor entity files.

This also fixes the path to the vendor script, which was given
relatively before, which obviously won't work when running from
inside a deeper subdirectory.

To speed things up, pass the --vendor option to getinfo, so we
do not have to find out the current vendor in getinfo all over
again.

Gbp-Dch: ignore