CMake: Translations: Don't rebuild .mo for line number changes
If only the line numbers changed in a file without any of the
translatable strings changing, the .po and .mo files were
rebuilt, making building simple code changes somewhat annoying.

We can work around this by passing --add-location=file to msgcomm
when we are creating the temporary .pot file used for building
the .mo files.