CMake: Handle Berkeley DB on FreeBSD
The BSD systems still ship their own db.h with a historical
BSD implementation, which is preferred by CMake, as it searches
its default path first. We thus have to disable the DEFAULT_PATH
for the search, unfortunately. We also need to pass the correct
include directory to the target.

Furthermore, on FreeBSD the library is called db-<VERSION>, so
let's add db-5 to the list of allowed names.

Gbp-Dch: ignore