test: Assert multi-arch in the chroot
The host system might not have a dpkg installed, which makes
dpkg fail with:

  dpkg not recorded as installed, cannot check for multi-arch support!

That's entirely useless of course. We want to know if dpkg could
support multi-arch in our chroot, so we pseudo-install dpkg into
the chroot and pretend it's version is one version higher than
the minimum dpkg version, so dpkg --assert-multi-arch works on
recent dpkgs.

Gbp-Dch: ignore