improve arch-unqualified dpkg-progress parsing
Our old idea was to look for the first package which would be "touched"
and take this as the package dpkg is talking about, but that is
incorrect in complicated situations like a package upgraded to/from
multiple M-A:same siblings installed.

As we us the progress report to decide what is still needed we have to
be reasonabily right about the package dpkg is talking about, so we jump
to quite a few loops to get it.

(cherry picked from commit 4b10240cca0dc0a4e82e42959545d2ae7e622d29)