avoid validate/delete/load race in cache generation
Keeping the Fd of the cache file we have validated around to later load
it into the mmap ensures not only that we load the same file (which
wouldn't really be a problem in practice), but that this file also still
exists and wasn't deleted e.g. by a 'apt clean' call run in parallel.

(cherry picked from commit 06606f073210fe3902fe92d5ff77fa1ab621b972)