Allow the daily script to be run in two phases
This adds an argument to the script which may be update, install,
or empty. In the update cases, downloads are performed. In the
install case, installs are performed. If empty, both are run.

Gbp-Dch: ignore
(cherry picked from commit 007b22ee6fd33d9b03637577f8401a43c5e7c90c)