apt.systemd.daily: Use unattend-ugrade --download-only if available
Using dry-run as in the previous commit is not really correct, as
it logs dpkg debugging output too. So, let's assume unattended-upgrade
gets a --download-only option and use that if it is available.

This lets us add the downloading part to unattended-upgrades later
on, without requiring versioned dependencies between the two.

Closes: #863859