fail instead of warn on insecure repositories in apt-get
The exception was made to give (script) users a one-release grace period
to adapt their setup to deal with apt enforcing signing of repositories.
As we are now at the start of a new release cycle its as good a time as
any to lift it now.

Removes-Exception: 952ee63b0af14a534c0aca00c11d1a99be6b22b2