send weak-only hashes to methods
Weak hashes like filesize can be used by methods for basic checks and
early refusals even if we can't use them for hard security proposes.

Normal apt operations are not affected by this as they fail if no strong
hash is available, but if apt is forced to work with weak-only files or
e.g. in apt-helper context it can have benefits as weak is better than
no hash for the methods.