apt.daily: remove unused dbus signal for apt update
The signal was introduced with the introduction of the script itself,
but seems to have never got any user as all references to it I can find
are references to other code appearing in the cronjob.

It is also the wrong place nowadays as the cronjob is just one place an
update can be triggered by, so if notifications about an update being
run are desired it is better to use a hook which will be called by all
update calls (script, cron, user, …).

Removing this code solves also the problem of improving the check to
avoid running into problems with security systems like SELinux.

References: 0c1326826fd23ce859db8e923c37b7199c6da2c8
Closes: 849636