allow the apt/lists/auxfiles/ directory to be missing
apt 1.6~alpha6 introduced aux requests to revamp the implementation of
a-t-mirror. This already included the potential of running as non-root,
but the detection wasn't complete resulting in errors or could produce
spurious warnings along the way if the directory didn't exist yet.

References: ef9677831f62a1554a888ebc7b162517d7881116
Closes: 887624