Don't use invalid iterator in Fallback-Of handling
cppcheck reports: (error) Iterator 't' used after element has been erased.

The loop is actually fashioned to deal with this (not in the most
efficient way, but in simplest and speed isn't really a concern here)
IF this codepath had a "break" at the end… so I added one.

Note that the tests aren't failing before (and hopefully after) the
change as the undefined behavior we encounter is too stable.

Thanks: David Binderman for reporting