Check new Release file passes its own Signed-By setting
A Release file can include a Signed-By header limiting which key can
sign the next Release file. It is hardly a security feature to require
the Release file containing this header to pass this test as well, but
it can prevent mistakes (like stale field contents) it would be harder
to recover from for a user. Not much by itself, but in the following
commit we will add another source of Signed-By data.