debListParser: Avoid native arch lookup in ParseDepends
We called low-level ParseDepends without an architecture each time,
which means each call looked up the native architecture. Store the
native architecture in the class and use that when calling low-level
ParseDepends from the high-level ParseDepends().

This improves performance for a cache build from 2.7 to 2.5 seconds
for me.

Also avoid a call when stripping multiarch, as the native architecture
is passed in.