Fix crashes in apt search, cache file
When BuildDepCaches() is called before BuildCaches() we end up
with dereferencing a null pointer cache in the depcache constructor.

Furthermore, in private-search, we check that Cache == NULL, but
only after constructing the policy and records, which does not
work reliably, because the records construction accesses the cache
as well.

LP: #1815187
Test-Case: ./build/bin/apt search -o RootDir=/dev/null foo