prepare-release: Add merge-translations command
This command allows merging translations from another branch. This
is to be used strictly downwards, that is, translations should be
merged from newest to oldest branch, without skipping any active
branch, to ensure the best result.

For example, consider branches 1 and 2 have the strings "a" and "b"
to translate, but master only has the strings "a" and "c". Now, 2
has more translations for "b" than 1. By merging master into 2, and
2 into 1, 1 also gets any translations for "b" from 2.