Introduce apt satisfy and apt-get satisfy
Allow to satisfy dependency strings supplied on
the command line, optionally prefixed with
"Conflicts:" to satisfy them like Conflicts.

Build profiles and architecture restriction lists,
as used in build dependencies, are supported as
well.

Compared to build-dep, build-essential is not
installed automatically, and installing of recommended
packages follows the global default, which defaults
to yes.

Closes: #275379
See merge request apt-team/apt!63