satisfy: Fix segmentation fault when called with empty argument
apt satisfy "" caused a segmentation fault because we were iterating
over the characters, checking if the next character was the end of
the string; when it could also be the current character.

Instead, check whether the next character is before the end of
the string, rather than identical to the end.