Avoid #define _error, use anonymous C++ struct instead
Replace the macro with an anonymous struct that provides an
inline operator->() returning the _error pointer.

This change is ABI compatible, and the inline macro is not
exported. We should consider if we want to avoid the function
call and directly export the thread_local variable instead,
when we do break ABI.

Closes: #948338