apt-mark: don't lie about successful marks
This commit fixes an issue where apt-mark would say it had made a change
before actually making the change. For example, when running as a user
without permission to write to extended_states, the package is not
marked but apt-mark claims it is:

    ~ % apt-mark manual rxvt-unicode
    rxvt-unicode set to manually installed.
    E: Could not create [...snip...] (13: Permission denied)
    E: Failed to write temporary StateFile /var/lib/apt/extended_states

This commit moves reporting of "[package] set to [manually |
automatically] installed" after saving extended_states and confirming it
was successful.